requirements.txt
opentargets
more-itertools
bpython
functional
pyyaml
addict
